test: isolate tier-3 live suites in TBDDaemonLiveTests + split CI into fast/quiet passes
Stage 0 of the test-hardening design (slice A). Kills the scheduler-contention
flake class at the source: live-tmux suites, deadline-racing subprocess suites,
and the replay firehose no longer share a starved runner with the 3859
deterministic tests that tick alongside them.

- New `TBDDaemonLiveTests` test target; 17 files moved plus 5 real-tmux tests
  extracted out of WorktreeLifecycleTests (30 dryRun tests stay parallel).
- test.yml runs two sequential steps in ONE job (the account caps macOS jobs at
  5 concurrent, so a second job is not an option): a fast `--parallel -j 2 --skip`
  pass, then a serial `--filter` quiet pass with `timeout-minutes: 15`.
- BOTH steps count-guard. `swift test --filter/--skip` exits GREEN on zero
  matches, so a renamed target could otherwise fall out of both passes and CI
  would report full coverage it never ran.
- Assertion-hygiene rules (design §6) and the tier taxonomy land in Tests/CLAUDE.md.

Verified: 3859 (fast) + 54 (quiet) == 3913 (unfiltered), all green.
