+++ Tests/TBDDaemonTests/ControlModeSettingsRPCTests.swift
@@ -1,5 +1,6 @@
 import Darwin
 import Foundation
+import TestSupport
 import Testing
 @testable import TBDDaemonLib
 @testable import TBDShared
